# Conflicts:
#	CMakeLists.txt
#	tests/algorithm/multiway_merge_benchmark.cpp
#	tests/algorithm/multiway_merge_test.cpp
#	tests/algorithm/random_bipartition_shuffle.cpp
#	tests/algorithm_test.cpp
#	tests/backtrace_test.cpp
#	tests/cmdline_parser_example.cpp
#	tests/cmdline_parser_test.cpp
#	tests/container/btree_speedtest.cpp
#	tests/container/btree_test.cpp
#	tests/container/d_ary_heap_speedtest.cpp
#	tests/container/d_ary_heap_test.cpp
#	tests/container/loser_tree_test.cpp
#	tests/container/lru_cache_test.cpp
#	tests/container/radix_heap_test.cpp
#	tests/container/ring_buffer_test.cpp
#	tests/container/simple_vector_test.cpp
#	tests/container/splay_tree_test.cpp
#	tests/container/string_view_test.cpp
#	tests/counting_ptr_test.cpp
#	tests/delegate_test.cpp
#	tests/deprecated_test.cpp
#	tests/die_test.cpp
#	tests/digest_test.cpp
#	tests/logger_test.cpp
#	tests/math/aggregate_test.cpp
#	tests/math/polynomial_regression_test.cpp
#	tests/math_test.cpp
#	tests/meta/call_for_test.cpp
#	tests/meta/fold_test.cpp
#	tests/meta/function_chain_test.cpp
#	tests/meta/function_stack_test.cpp
#	tests/meta/has_member_test.cpp
#	tests/meta/log2_test.cpp
#	tests/meta/vmap_for_test.cpp
#	tests/multi_timer_test.cpp
#	tests/semaphore_test.cpp
#	tests/siphash_test.cpp
#	tests/sort_networks_test.cpp
#	tests/sort_parallel_mergesort_test.cpp
#	tests/sort_strings_example.cpp
#	tests/sort_strings_parallel_test.cpp
#	tests/sort_strings_test.cpp
#	tests/stack_allocator_test.cpp
#	tests/string_test.cpp
#	tests/thread_barrier_test.cpp
#	tests/thread_pool_test.cpp
#	tlx/container/btree.hpp
#	tlx/container/d_ary_addressable_int_heap.hpp
#	tlx/sort/strings/string_set.hpp

+++ tests/meta/vmap_for_test.cpp
@@ -8,34 +8,41 @@
  * All rights reserved. Published under the Boost Software License, Version 1.0
  ******************************************************************************/
 
-#include <cmath>
-#include <cstddef>
-#include <iostream>
-#include <sstream>
-
 #include <tlx/die.hpp>
 #include <tlx/meta/vmap_for_range.hpp>
 #include <tlx/meta/vmap_foreach.hpp>
 #include <tlx/meta/vmap_foreach_tuple.hpp>
 #include <tlx/meta/vmap_foreach_tuple_with_index.hpp>
 #include <tlx/meta/vmap_foreach_with_index.hpp>
+#include <cmath>
+#include <cstddef>
+#include <iostream>
+#include <sstream>
+#include <string>
+#include <tuple>
 
 /******************************************************************************/
 // vmap_foreach
 
-struct SimpleMapFunctor {
-    explicit SimpleMapFunctor(std::ostream& os) : os_(os) { }
+struct SimpleMapFunctor
+{
+    explicit SimpleMapFunctor(std::ostream& os) : os_(os)
+    {
+    }
+
     template <typename Arg>
-    auto operator () (const Arg& a) const {
+    auto operator()(const Arg& a) const
+    {
         os_ << a << '\n';
         return a + 1;
     }
+
     std::ostream& os_;
 };
 
 template <typename... Args>
-void test_vmap_foreach_run(std::ostream& os, const Args& ... args) {
-
+void test_vmap_foreach_run(std::ostream& os, const Args&... args)
+{
     auto r1 = tlx::vmap_foreach(
         [&os](auto a) {
             os << a << '\n';
@@ -43,25 +50,24 @@ void test_vmap_foreach_run(std::ostream& os, const Args& ... args) {
         },
         args...);
 
-    die_unequal(std::tuple_size<decltype(r1)>::value, 3u);
+    die_unequal(std::tuple_size<decltype(r1)>::value, 3U);
     die_unequal(std::get<0>(r1), 43);
     die_unequal(std::get<1>(r1), 6.0);
     die_unequal(std::get<2>(r1), std::string("ello"));
 
     auto r2 = tlx::vmap_foreach(SimpleMapFunctor(os), args...);
 
-    die_unequal(std::tuple_size<decltype(r2)>::value, 3u);
+    die_unequal(std::tuple_size<decltype(r2)>::value, 3U);
     die_unequal(std::get<0>(r2), 43);
     die_unequal(std::get<1>(r2), 6.0);
     die_unequal(std::get<2>(r2), std::string("ello"));
 }
 
-static void test_vmap_foreach() {
-
+static void test_vmap_foreach()
+{
     std::ostringstream oss;
 
-    test_vmap_foreach_run(
-        oss, static_cast<int>(42), static_cast<double>(5), "hello");
+    test_vmap_foreach_run(oss, 42, static_cast<double>(5), "hello");
 
     die_unequal("42\n5\nhello\n42\n5\nhello\n", oss.str());
 }
@@ -70,8 +76,8 @@ static void test_vmap_foreach() {
 // vmap_foreach_tuple
 
 template <typename... Args>
-void test_vmap_foreach_tuple_run(std::ostream& os, const Args& ... args) {
-
+void test_vmap_foreach_tuple_run(std::ostream& os, const Args&... args)
+{
     auto my_tuple = std::make_tuple(args...);
 
     auto r1 = tlx::vmap_foreach_tuple(
@@ -81,25 +87,24 @@ void test_vmap_foreach_tuple_run(std::ostream& os, const Args& ... args) {
         },
         my_tuple);
 
-    die_unequal(std::tuple_size<decltype(r1)>::value, 3u);
+    die_unequal(std::tuple_size<decltype(r1)>::value, 3U);
     die_unequal(std::get<0>(r1), 43);
     die_unequal(std::get<1>(r1), 6.0);
     die_unequal(std::get<2>(r1), std::string("ello"));
 
     auto r2 = tlx::vmap_foreach_tuple(SimpleMapFunctor(os), my_tuple);
 
-    die_unequal(std::tuple_size<decltype(r2)>::value, 3u);
+    die_unequal(std::tuple_size<decltype(r2)>::value, 3U);
     die_unequal(std::get<0>(r2), 43);
     die_unequal(std::get<1>(r2), 6.0);
     die_unequal(std::get<2>(r2), std::string("ello"));
 }
 
-static void test_vmap_foreach_tuple() {
-
+static void test_vmap_foreach_tuple()
+{
     std::ostringstream oss;
 
-    test_vmap_foreach_tuple_run(
-        oss, static_cast<int>(42), static_cast<double>(5), "hello");
+    test_vmap_foreach_tuple_run(oss, 42, static_cast<double>(5), "hello");
 
     die_unequal("42\n5\nhello\n42\n5\nhello\n", oss.str());
 }
@@ -107,19 +112,25 @@ static void test_vmap_foreach_tuple() {
 /******************************************************************************/
 // vmap_foreach_with_index
 
-struct SimpleMapFunctorWithIndex {
-    explicit SimpleMapFunctorWithIndex(std::ostream& os) : os_(os) { }
+struct SimpleMapFunctorWithIndex
+{
+    explicit SimpleMapFunctorWithIndex(std::ostream& os) : os_(os)
+    {
+    }
+
     template <typename Index, typename Arg>
-    auto operator () (const Index& i, const Arg& a) const {
+    auto operator()(const Index& i, const Arg& a) const
+    {
         os_ << i << ' ' << a << '\n';
         return a + 1;
     }
+
     std::ostream& os_;
 };
 
 template <typename... Args>
-void test_vmap_foreach_with_index_run(std::ostream& os, const Args& ... args) {
-
+void test_vmap_foreach_with_index_run(std::ostream& os, const Args&... args)
+{
     // a variadic tuple to get indexes from using the compile-time index param
     auto my_tuple = std::make_tuple(1, 'x', "abc");
 
@@ -131,37 +142,38 @@ void test_vmap_foreach_with_index_run(std::ostream& os, const Args& ... args) {
         },
         args...);
 
-    die_unequal(std::tuple_size<decltype(r1)>::value, 3u);
+    die_unequal(std::tuple_size<decltype(r1)>::value, 3U);
     die_unequal(std::get<0>(r1), 43);
     die_unequal(std::get<1>(r1), 6.0);
     die_unequal(std::get<2>(r1), std::string("ello"));
 
-    auto r2 = tlx::vmap_foreach_with_index(
-        SimpleMapFunctorWithIndex(os), args...);
+    auto r2 =
+        tlx::vmap_foreach_with_index(SimpleMapFunctorWithIndex(os), args...);
 
-    die_unequal(std::tuple_size<decltype(r2)>::value, 3u);
+    die_unequal(std::tuple_size<decltype(r2)>::value, 3U);
     die_unequal(std::get<0>(r2), 43);
     die_unequal(std::get<1>(r2), 6.0);
     die_unequal(std::get<2>(r2), std::string("ello"));
 }
 
-static void test_vmap_foreach_with_index() {
-
+static void test_vmap_foreach_with_index()
+{
     std::ostringstream oss;
 
-    test_vmap_foreach_with_index_run(
-        oss, static_cast<int>(42), static_cast<double>(5), "hello");
+    test_vmap_foreach_with_index_run(oss, 42, static_cast<double>(5), "hello");
 
     die_unequal("0 42 1\n1 5 x\n2 hello abc\n"
-                "0 42\n1 5\n2 hello\n", oss.str());
+                "0 42\n1 5\n2 hello\n",
+                oss.str());
 }
 
 /******************************************************************************/
 // vmap_foreach_tuple_with_index
 
 template <typename Tuple>
-void test_vmap_foreach_tuple_with_index_run(std::ostream& os, const Tuple& tuple) {
-
+void test_vmap_foreach_tuple_with_index_run(std::ostream& os,
+                                            const Tuple& tuple)
+{
     // a variadic tuple to get indexes from using the compile-time index param
     auto my_tuple = std::make_tuple(1, 'x', "abc");
 
@@ -173,58 +185,56 @@ void test_vmap_foreach_tuple_with_index_run(std::ostream& os, const Tuple& tuple
         },
         tuple);
 
-    die_unequal(std::tuple_size<decltype(r1)>::value, 3u);
+    die_unequal(std::tuple_size<decltype(r1)>::value, 3U);
     die_unequal(std::get<0>(r1), 43);
     die_unequal(std::get<1>(r1), 6.0);
     die_unequal(std::get<2>(r1), std::string("ello"));
 
-    auto r2 = tlx::vmap_foreach_tuple_with_index(
-        SimpleMapFunctorWithIndex(os), tuple);
+    auto r2 = tlx::vmap_foreach_tuple_with_index(SimpleMapFunctorWithIndex(os),
+                                                 tuple);
 
-    die_unequal(std::tuple_size<decltype(r2)>::value, 3u);
+    die_unequal(std::tuple_size<decltype(r2)>::value, 3U);
     die_unequal(std::get<0>(r2), 43);
     die_unequal(std::get<1>(r2), 6.0);
     die_unequal(std::get<2>(r2), std::string("ello"));
 }
 
-static void test_vmap_foreach_tuple_with_index() {
-
+static void test_vmap_foreach_tuple_with_index()
+{
     std::ostringstream oss;
 
     test_vmap_foreach_tuple_with_index_run(
-        oss,
-        std::make_tuple(static_cast<int>(42), static_cast<double>(5), "hello"));
+        oss, std::make_tuple(42, static_cast<double>(5), "hello"));
 
     die_unequal("0 42 1\n1 5 x\n2 hello abc\n"
-                "0 42\n1 5\n2 hello\n", oss.str());
+                "0 42\n1 5\n2 hello\n",
+                oss.str());
 }
 
 /******************************************************************************/
 // vmap_for_range
 
-static void test_vmap_for_range() {
-
+static void test_vmap_for_range()
+{
     std::ostringstream oss;
 
-    auto r1 = tlx::vmap_for_range<6>(
-        [&](size_t index) {
-            oss << index << ' ';
-            return std::to_string(index + 42);
-        });
+    auto r1 = tlx::vmap_for_range<6>([&](size_t index) {
+        oss << index << ' ';
+        return std::to_string(index + 42);
+    });
 
-    die_unequal(std::tuple_size<decltype(r1)>::value, 6u);
+    die_unequal(std::tuple_size<decltype(r1)>::value, 6U);
     die_unequal(std::get<0>(r1), "42");
     die_unequal(std::get<1>(r1), "43");
 
-    auto r2 = tlx::vmap_for_range<4, 8>(
-        [&](auto index) {
-            oss << index << ' ';
-            return index + 42;
-        });
+    auto r2 = tlx::vmap_for_range<4, 8>([&](auto index) {
+        oss << index << ' ';
+        return index + 42;
+    });
 
-    die_unequal(std::tuple_size<decltype(r2)>::value, 4u);
-    die_unequal(std::get<0>(r2), 46u);
-    die_unequal(std::get<1>(r2), 47u);
+    die_unequal(std::tuple_size<decltype(r2)>::value, 4U);
+    die_unequal(std::get<0>(r2), 46U);
+    die_unequal(std::get<1>(r2), 47U);
 
     die_unequal("0 1 2 3 4 5 4 5 6 7 ", oss.str());
 }
